chore: move inputTypes to types folder
